Fix up E, K, L
With E and K, the last line had the wrong amount of whitespace
in it.

L was missing its last line completely. This would have created
a lookup error, if not for the fact that the last line was never
printed.

I think this is an oversight, and the last line should be printed,
otherwise we wouldn't get the "tail" of Q. So I went and added
printing of the last line.